requirements.txt
agg==0.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
attrs==22.2.0
Babel==2.11.0
beautifulsoup4==4.11.1
bleach==5.0.1
camelcase==0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-ratings==0.3.7
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
dnspython==2.2.1
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
fqdn==1.5.1
httpie==3.2.1
idna==3.4
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_server==2.0.5
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.17.0
kiwisolver==1.4.4
lib==4.0.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdurl==0.1.2
mistune==2.0.4
mplstyle==0.1.2
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
packages==0.1.0
pandas==1.5.2
pandocfilters==1.5.0
Pillow==9.3.0
platformdirs==2.6.2
prometheus-client==0.15.0
protobuf==3.20.1
psycopg2==2.9.5
pybind11==2.10.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.7
pywinpty==2.0.9
PyYAML==6.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
somepackages==0.1.3
soupsieve==2.3.2.post1
speedtest-cli==2.1.3
sqlparse==0.4.3
stats==0.1.2a0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.13
userprovided==0.9.4
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
